Upgrade the return type of GetLength from int to int64.
The size of file can be larger than 2G, so int may not have enough range to
return it. Considering the callback already takes int64, it should be ok to use
it for the return type too.

BUG=127129
TEST=Ran unit_tests


Review URL: https://chromiumcodereview.appspot.com/13113003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@190942 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 116a31209b25eaca2ef6cfaaf371c37a7f86dcb2

+++ chrome/browser/chromeos/drive/drive_file_stream_reader.cc
@@ -21,7 +21,7 @@ int DriveFileStreamReader::Read(net::IOBuffer* buf, int buf_len,
   return 0;
 }
 
-int DriveFileStreamReader::GetLength(
+int64 DriveFileStreamReader::GetLength(
     const net::Int64CompletionCallback& callback) {
   // TODO(hidehiko): Implement this.
   NOTIMPLEMENTED();
